daemon: move maskCredentials to config package
This allows the utility to be used in other places.
